Refactor security and database configuration to use settings from `app/config.py`
**Commit Message:**
- Moved secret keys and database URLs to `app/config.py` for better management.
- Updated imports in `auth_utils.py`, `database.py`, and `spell_effects.py` to use the new settings.

+++ app/spell_effects.py
@@ -167,14 +167,14 @@ def effect_magic_push(hero,session):
 
     if random.random() > 0.5:
         
-        damage = hero_damage *2
+        damage = int(hero_damage *2)
         log.append(f"You used magic push on {monster.name} it fell and took {damage} critical damage.")
     else:
         
-        damage = hero_damage
+        damage = int(hero_damage)
         log.append(f"You used magic push on {monster.name} it took {damage} damage.")
 
-    monster.current_hp -= int(damage)
+    monster.current_hp -= damage
     
     return log
